requirements.txt
tqdm==4.50.0
pyyaml==5.4.1
scipy==1.0.1
matplotlib==3.3.2
Pillow==7.1.2
scikit-image==0.17.2
scikit-learn==0.23.2
pandas==1.1.2
seaborn==0.11.0
tables==3.6.1
numba==0.51.2
opencv-python==4.4.0.44
torch==1.4.0
torchvision==0.5.0
tensorboard==2.4.1
